+++ src/units.rs
@@ -1,5 +1,5 @@
 /// A height or number of blocks
-pub type Height = usize;
+pub type Height = u64;
 
 /// A number of zatoshi
-pub type Zat = usize;
+pub type Zat = u64;
